Refactor contingencyFlow zoid component to support reloads
