+++ src/component.js
@@ -10,6 +10,7 @@ import { destroy as zoidDestroy } from 'zoid/src';
 import btClient from '../vendor/braintree-web/client';
 import hostedFields from '../vendor/braintree-web/hosted-fields';
 
+import { getContingencyFlowComponent } from './zoid/contingency-flow';
 import contingencyFlow from './contingency-flow';
 import type { HostedFieldsHandler } from './types';
 
@@ -78,7 +79,7 @@ function createSubmitHandler (hostedFieldsInstance, orderIdFunction) : Function 
         const url = `${ err.links.find(link => link.rel === '3ds-contingency-resolution').href  }`;
 
         logger.info('HOSTEDFIELDS_3DS');
-        return contingencyFlow.start(url);
+        return contingencyFlow.start(getContingencyFlowComponent(), url);
       }).then((payload) => {
         // does contingency flow give a payload?
         logger.track({
@@ -256,6 +257,9 @@ export function setupHostedFields() : Function {
   if (isChildWindow()) {
     return;
   }
+
+  // initialize the contingency flow zoid component
+  getContingencyFlowComponent();
 
   const merchantId = getMerchantID();
   const originalFundingEligibility = getFundingEligibility();
