+++ src/zoid/contingency-flow/index.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export * from './component';
