+++ src/zoid/contingency-flow/props.js
@@ -1,0 +1,9 @@
+/* @flow */
+
+export type ContingencyFlowProps = {|
+    action : string,
+    cart_id : string,
+    flow : string,
+    xcomponent : string,
+    onContingencyResult : (err : mixed, result : Object) => void
+|};
