+++ test/component.js
@@ -12,6 +12,7 @@ import btClient from '../vendor/braintree-web/client';
 import hostedFields from '../vendor/braintree-web/hosted-fields';
 import { HostedFields, setupHostedFields } from '../src/index';
 import contingencyFlow from '../src/contingency-flow';
+import { getContingencyFlowComponent } from '../src/zoid/contingency-flow';
 
 import rejectIfResolves from './utils/reject-if-resolves';
 
@@ -452,11 +453,12 @@ describe('hosted-fields-component', () => {
       td.when(fakeHostedFieldsInstance.tokenize(td.matchers.isA(Object)))
         .thenReject(error);
 
-      td.when(contingencyFlow.start(expectedUrl)).thenResolve(threeDSResult);
+      td.when(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl)).thenResolve(threeDSResult);
 
       return HostedFields.render(renderOptions, '#button').then((handler) => {
         return handler.submit().then((actual) => {
-          td.verify(contingencyFlowStart(expectedUrl));
+
+          td.verify(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl));
           assert.equal(actual.liabilityShifted, true);
         });
       });
@@ -495,11 +497,11 @@ describe('hosted-fields-component', () => {
       td.when(fakeHostedFieldsInstance.tokenize(td.matchers.isA(Object)))
         .thenReject(error);
 
-      td.when(contingencyFlow.start(expectedUrl)).thenReject();
+      td.when(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl)).thenReject();
 
       return HostedFields.render(renderOptions, '#button').then((handler) => {
         return handler.submit().then(rejectIfResolves).catch(() => {
-          td.verify(contingencyFlowStart(expectedUrl));
+          td.verify(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl));
         });
       });
     });
@@ -544,11 +546,11 @@ describe('hosted-fields-component', () => {
       td.when(fakeHostedFieldsInstance.tokenize(td.matchers.isA(Object)))
         .thenReject(error);
 
-      td.when(contingencyFlow.start(expectedUrl)).thenResolve(threeDSResult);
+      td.when(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl)).thenResolve(threeDSResult);
 
       return HostedFields.render(renderOptions, '#button').then((handler) => {
         return handler.submit().then((actual) => {
-          td.verify(contingencyFlowStart(expectedUrl));
+          td.verify(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl));
           assert.equal(actual.liabilityShifted, true);
           assert.equal(actual.authenticationStatus, threeDSResult.status);
           assert.equal(actual.authenticationReason, threeDSResult.authentication_status_reason);
@@ -596,11 +598,11 @@ describe('hosted-fields-component', () => {
       td.when(fakeHostedFieldsInstance.tokenize(td.matchers.isA(Object)))
         .thenReject(error);
 
-      td.when(contingencyFlow.start(expectedUrl)).thenResolve(threeDSResult);
+      td.when(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl)).thenResolve(threeDSResult);
 
       return HostedFields.render(renderOptions, '#button').then((handler) => {
         return handler.submit().then((actual) => {
-          td.verify(contingencyFlowStart(expectedUrl));
+          td.verify(contingencyFlowStart(getContingencyFlowComponent(), expectedUrl));
           assert.equal(actual.liabilityShifted, false);
           assert.equal(actual.authenticationStatus, threeDSResult.status);
           assert.equal(actual.authenticationReason, threeDSResult.authentication_status_reason);
